+++ R/pat_load.R
@@ -1,6 +1,6 @@
 #' @export
 #' @importFrom rlang .data
-#' @importFrom dplyr filter
+#' @importFrom MazamaCoreUtils logger.debug
 #' 
 #' @title Load Purple Air Timeseries data
 #' 
@@ -8,28 +8,33 @@
 #' the given time interval. Archived data for SCAQMD sensors go back to 
 #' January, 2018.
 #' 
-#' The \code{startdate} and \code{enddate} can be in the following format:
+#' Dates can be anything that is understood by 
+#' \code{lubrdiate::parse_date_time()} including either of the following 
+#' recommended formats:
 #' 
 #' \itemize{
 #' \item{\code{"YYYYmmdd"}}
+#' \item{\code{"YYYY-mm-dd"}}
 #' }
 #' 
-#' By default, the current month is loaded.
+#' By default, the current week is loaded.
 #'
 #' @param label Purple Air 'label'
-#' @param startdate A starting date in ymd order.
-#' @param enddate An ending date in ymd order
+#' @param startdate desired start datetime (ISO 8601)
+#' @param enddate desired end datetime (ISO 8601)
+#' @param days number of days of data to include
 #' @param timezone Timezone used to interpret datestamp.
 #' @param baseUrl Base URL for synoptic data.
 #' 
 #' @return An object of class "pa_timeseries".
 #' 
 #' @seealso \link{pat_loadMonth}
+#' @seealso \link{.dateRange}
 #' 
 #' @examples
 #' \dontrun{
-#' pat_load("SCNP_20", 20190411, 20190521)
-#'   pat_multiplot()
+#' pat <- pat_load("SCNP_20", 20190411, 20190521)
+#' pat_multiplot(pat)
 #' }
 
 pat_load <- 
@@ -37,48 +42,50 @@ pat_load <-
     label = NULL, 
     startdate = NULL, 
     enddate = NULL, 
+    days = 7, 
     timezone = "America/Los_Angeles",
     baseUrl = "http://smoke.mazamascience.com/data/PurpleAir/pat"
     ) {
      
     logger.debug("----- pat_load() -----")
     
+    # Validate parameters ------------------------------------------------------
+    
     if ( is.null(label) ) 
       stop("Required parameter 'label' is missing.")
     
-    if (is.null(startdate) || is.null(enddate) ) 
-      stop("Required start and/or end date is missing")
+    dateRange <- .dateRange(startdate, enddate, days, timezone)
     
-    ymdBounds <- 
-      c(startdate, enddate) %>% 
-      lubridate::ymd(tz = timezone)
+    # Asssemble monthly archive files ------------------------------------------
     
-    ymInt <- 
+    datestamps <-
       seq(
-        ymdBounds[1], 
-        ymdBounds[2], 
+        dateRange[1],
+        dateRange[2],
         by = "months"
-      ) %>% 
+      ) %>%
       strftime(format = "%Y%m")
     
     patList <- list()
     
-    for ( i in ymInt ) { 
+    for ( datestamp in datestamps ) { 
       
-      patList[[i]] <- 
+      patList[[datestamp]] <- 
         pat_loadMonth(
           label = label, 
-          datestamp = i, 
+          datestamp = datestamp, 
           timezone = timezone,
           baseUrl = baseUrl
         )
       
     } 
     
+    # Return -------------------------------------------------------------------
+    
     patObj <- 
       patList %>% 
       pat_join() %>% 
-      pat_filterDate(ymdBounds[1], ymdBounds[2])
+      pat_filterDate(dateRange[1], dateRange[2])
    
     return(patObj)